pyproject.toml
[tool.poetry]
name = "mcp-repocache"
version = "0.1.0"
description = "MCP Repository Cache - Documentation aggregation and indexing system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "mcp"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
gitpython = "^3.1.40"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
black = "^25.10.0"
flake8 = "^6.0.0"
httpx = "^0.25.0"  # For FastAPI test client
pylint = "^4.0.4"
safety = "^3.7.0"
bandit = "^1.9.2"
mypy = "^1.19.1"
ruff = "^0.14.10"
hypothesis = "^6.148.8"
pytest-benchmark = "^5.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
index-docs = "scripts.index_docs:main"
run-server = "scripts.run_server:main"
test = "pytest"
test-cov = "pytest --cov=mcp --cov=scripts --cov-report=html --cov-report=term"
safety-check = "safety check"
safety-review = "safety check --output text"

[tool.safety]
# Safety configuration for dependency vulnerability scanning
exclude = [
    "data/raw/*",
    "data/raw",
    ".git/*",
    ".venv/*",
    "__pycache__/*",
    "*.pyc"
]
output = "text"
continue-on-error = true
cache-expires = 3600
max-age = 24
